Automtic-screen-clicker
Program Description: Automatic Clicking and User Interaction

This Python program attempts to locate a specific image on the screen using the pyautogui module. When the image is found, the program clicks on the specified image and moves the mouse a certain distance. Then, it gives the user the option to continue the program or exit.

How to Use:

Before starting the program, create or choose an image file representing a specific image (deneme.png).
Place the created or selected deneme.png file in the same directory as the program file or in the file path specified in the code (C:\deneme.png).
Start the program.
The program will start searching for the specified image (deneme.png) on the screen.
If the specified image is found, the program will automatically click on the image and move the mouse a certain distance.
Then, the user will be prompted to press '1' to continue or '2' to exit.
Pressing '2' will exit the program.
Development Environment:

Developed using the Python programming language.
The pyautogui module is used for mouse and keyboard interactions.
Notes:

Programs like these, involving automatic clicking and user interaction, can be used for automation, especially for specific screen scenarios.
At the beginning of the program, it is expected to click on a specific image. The image should be in a specific location or state for the program to work correctly.
Execution Steps:

Place the program file in the C:\ directory.
Place the deneme.png file in the C:\ directory.
Convert it to an .exe file using PyInstaller and start using the program.